Make api tests pass.
