--- cdx_api_core/app/controllers/cdx_api_core/api/filters_controller.rb
@@ -1,12 +1,0 @@
-class Api::FiltersController < ApiController
-  respond_to :json
-  expose(:filters) { current_user.filters }
-
-  def index
-    respond_with filters
-  end
-
-  def show
-    render :edit
-  end
-end
