--- cdx_api_core/app/controllers/cdx_api_core/api/playground_controller.rb
@@ -1,9 +1,0 @@
-class Api::PlaygroundController <ApiController
-  def index
-    @devices = check_access(Device, READ_DEVICE)
-  end
-
-  def simulator
-    @devices = check_access(Device, READ_DEVICE)
-  end
-end
