+++ cdx_api_core/app/controllers/cdx_api_core/api_controller.rb
@@ -1,5 +1,5 @@
 module CdxApiCore
-  class ApiController < ApplicationController
+  class ApiController < ::ApplicationController
     include ApplicationHelper
     include Policy::Actions
