--- cdx_api_core/app/controllers/cdx_api_core/application_controller.rb
@@ -1,5 +1,0 @@
-module CdxApiCore
-  class ApplicationController < ActionController::Base
-
-  end
-end
