+++ cdx_api_core/app/controllers/cdx_api_core/filters_controller.rb
@@ -1,0 +1,14 @@
+module CdxApiCore
+  class FiltersController < CdxApiCore::ApiController
+    respond_to :json
+    expose(:filters) { current_user.filters }
+
+    def index
+      respond_with filters
+    end
+
+    def show
+      render :edit
+    end
+  end
+end
