+++ cdx_api_core/app/controllers/cdx_api_core/playground_controller.rb
@@ -1,0 +1,13 @@
+module CdxApiCore
+  class PlaygroundController < CdxApiCore::ApiController
+    layout 'api'
+
+    def index
+      @devices = check_access(Device, READ_DEVICE)
+    end
+
+    def simulator
+      @devices = check_access(Device, READ_DEVICE)
+    end
+  end
+end
