+++ cdx_api_core/app/views/cdx_api_core/encounters/index.csv.csvbuilder
