+++ cdx_api_core/app/views/cdx_api_core/playground/simulator.haml
