+++ cdx_api_core/app/views/layouts/api.haml
@@ -1,0 +1,12 @@
+!!!
+%html
+  %head
+    %title Connected Diagnostics Platform
+    = Gon::Base.render_data({})
+    = stylesheet_link_tag "application", :media => "all"
+    = javascript_include_tag "es5-shim" if Rails.env.test?
+    = javascript_include_tag "application"
+    = csrf_meta_tags
+  %body.devise{class: @body_class}
+    .wrapper
+      = content_for?(:cdp_content) ? yield(:cdp_content) : yield
