--- cdx_api_core/app/views/layouts/cdx_api_core/application.html.erb
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>CdxApiCore</title>
-  <%= stylesheet_link_tag    "cdx_api_core/application", media: "all" %>
-  <%= javascript_include_tag "cdx_api_core/application" %>
-  <%= csrf_meta_tags %>
-</head>
-<body>
-
-<%= yield %>
-
-</body>
-</html>
