+++ cdx_api_core/config/routes.rb
@@ -1,39 +1,37 @@
 CdxApiCore::Engine.routes.draw do
-  namespace :api, defaults: { format: 'json' } do
-    resources :activations, only: :create
-    resources :playground, only: :index, defaults: { format: 'html' } do
-      collection do
-        get :simulator
-      end
+  resources :activations, only: :create
+  resources :playground, only: :index, defaults: { format: 'html' } do
+    collection do
+      get :simulator
     end
-    match 'tests(.:format)' => "tests#index", via: [:get, :post]
-    resources :tests, only: [] do
-      collection do
-        get :schema
-      end
-      member do
-        get :pii
-      end
+  end
+  match 'tests(.:format)' => "tests#index", via: [:get, :post]
+  resources :tests, only: [] do
+    collection do
+      get :schema
     end
-    match 'encounters(.:format)' => "encounters#index", via: [:get, :post]
-    resources :encounters, only: [] do
-      collection do
-        get :schema
-      end
-      member do
-        get :pii
-      end
+    member do
+      get :pii
     end
-    resources :devices, only: [] do
-      resources :messages, only: [:create ], shallow: true
-      match 'tests' => "messages#create", via: :post # For backwards compatibility with Qiagen-Esequant-LR3
-      match 'demodata' => "messages#create_demo", via: :post
+  end
+  match 'encounters(.:format)' => "encounters#index", via: [:get, :post]
+  resources :encounters, only: [] do
+    collection do
+      get :schema
     end
-    resources :sites, only: :index
-    resources :institutions, only: :index
-    resources :filters, only: [:index, :show] do
-      resources :subscribers
+    member do
+      get :pii
     end
+  end
+  resources :devices, only: [] do
+    resources :messages, only: [:create ], shallow: true
+    match 'tests' => "messages#create", via: :post # For backwards compatibility with Qiagen-Esequant-LR3
+    match 'demodata' => "messages#create_demo", via: :post
+  end
+  resources :sites, only: :index
+  resources :institutions, only: :index
+  resources :filters, only: [:index, :show] do
     resources :subscribers
   end
+  resources :subscribers
 end
