+++ cdx_api_core/spec/controllers/cdx_api_core/activations_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe CdxApiCore::Api::ActivationsController do
+describe CdxApiCore::ActivationsController do
   let!(:device) { Device.make }
 
   shared_context :set_device_token do
