+++ cdx_api_core/spec/controllers/cdx_api_core/messages_controller_auth_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Api::MessagesController, elasticsearch: true, validate_manifest: false do
+describe CdxApiCore::MessagesController, elasticsearch: true, validate_manifest: false do
 
   let(:user) {User.make}
   let(:institution) {Institution.make user_id: user.id}
