+++ config/routes.rb
@@ -4,9 +4,8 @@ require 'sidekiq/cron/web'
 Rails.application.routes.draw do
   use_doorkeeper
 
-  mount CdxApiCore::Engine => "/api"
-
   #mount LetterOpenerWeb::Engine, at: '/letter_opener' if Rails.env == 'development'
+  mount CdxApiCore::Engine => "/api"
 
   if Rails.env.development?
     get 'join' => 'home#join'
